requirements.txt
appdirs==1.4.4
argcomplete==2.0.0
args==0.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
clint==0.5.1
colorama==0.4.5
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
processing-py==0.3.7
pyserial==3.5
python-escpos==2.2.0
pyusb==1.2.1
PyYAML==6.0
qrcode==7.3.1
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
